bugworld2 update - there are plants
The plants are created into an array list in world. The bugs arraylist
is now stored in world. For the life of me i cannot get the collision
detection to work properly in the population method. Still need to get
the created bugs to collision detect against plants also. Not even
started on obstacles. Had an idea for a generalObject class to store all
the onscreen elements in a stripped down format. Wasn't working, but
might revisit. Would make drawing different objects to the screen
easier. Hopefully.